Update readme & add sample rss file
In the switch to python3, something had to be configured in the
appengine console in order for some things to work on the backend. That
change is now documented as a part of starting a new podcast app.

I also added a sample rss file so that new podcast apps will be easier
to spin up.

I added a "Future improvements" section to the readme so that I can
somewhat track my ideas for enhancing the application.